Try Map of MutableDouble for CatalogUtils
CatalogUtils compute many signal parameters stored with double values.
They are passed to many methods arguments and modified in many places.
Leveraging Map references to group these double values
would help to simplify the code by avoiding long list of such parameters
and allowing their values to be modified directly inside intermediate
functions.
But Double is immutable and incompatible with that usage.
Apache proposes a MutableDouble alternative.
This is still a proof of concept just to test this implementation.

Of course, there is still the possibility to use double[] arrays.
But w/o key names to describe, the code reading will be harder.

Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: Iff3cf98d82cb53dc3cc89714b92a2e3bf77157fb