Allow the client to specify service port
The feature is mainly implemented in PceOpticalNode and
PceCalculation.

The rest of the classes are tests being updated since there is a new
constructor argument in PceOpticalNode.

Feature implementation overview

The feature uses the classes in package:
org.opendaylight.transportpce.pce.networkanalyzer.port

PceCalculation instantiates an instance of Preference using
PreferenceFactory and passing it on to the constructor in
PceOpticalNode.
PreferenceFactory in turn uses PathComputationRequestInput
to return either NoPreference or ClientPreference.
The class NoPreference ensures the outcome is the same as
before this feature was implemented.
PceOpticalNode uses ClientPreference.preferredPort(...) when
building a list of avilable ports. If a port is not in the
preferred list, it's treated as unavailable by PceOpticalNode.

APIImpact

POST /rests/operations/org-openroadm-service:service-create
{
  "input": {
...
    "port": {
      "port-device-name": "ROADM-B-SRG1",
      "port-name":        "SRG1-PP2-TXRX",
    }
...
  }
}

The above will result in port SRG1-PP2-TXRX on node ROADM-B-SRG1
being used in the service. All other ports on ROADM-B-SRG1
will be treated as "unavailable".

JIRA: TRNSPRTPCE-176
Change-Id: I453a92fda2977ac86c5c5ff47e1a7f6cd322334d
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>