Specify SRG for ROADM-to-ROADM service
Currently, when requesting a ROADM-to-ROADM service (from SRG PP to
SRG PP) using the OpenROADM service API, the PCE just selects the first
available SRG and PP on the end nodes. According to the
org-openroadm-common-service-types yang module, the endpoint port
attribute in the service-create RPC input uniquely identifies the
termination point (port-device-name + port-circuit-pack-name +
port-name).

This change makes it possible to specify SRG using port-device-name
(note that it was already possible to specify Xponder endpoint using
port-device-name when service-format is OTU).

JIRA: TRNSPRTPCE-176
Signed-off-by: Jonas Mårtensson <jonas.martensson@smartoptics.com>
Change-Id: Ie09b23da30ca5c2062f49c54e4c809e824414d95