New observer package subscribing for error msg
The class EventSubscriber can be used to track events of various
severity. The class is intended as a way to collect error messages
for later use. The class contains convenient methods extracting the
first or last message from a given type (error, warning etc).

The class is particularly useful in the case of an iteration where it's
not uncommon for errors to occur during the iteration, although at some
point in the iteration at least one item should complete without error.

Change-Id: I9210afca69f516a7ef3788e854b1af4d8bbcc450
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>