Deleting openroadm node after controller restart
When creating an openroadm node, the internal node topology is put into
variable topologyShardMountedDevice. Later, when deleting the node, the
topology is retrieved from this variable, and nodes and links are
deleted from openroadm topology. The problem is that the
topologyShardMountedDevice variable is cleared when TransportPCE is
restarted for whatever reason, causing node deletion to fail.

Change to reading the openroadm topology from the datastore instead and
find nodes and links to delete.

JIRA: TRNSPRTPCE-797
Change-Id: I1d605eeb0c9a72701e0935580f02123bd421be7f
Signed-off-by: Jonas Mårtensson <jonas.martensson@smartoptics.com>