New Tapi package dealing with frequency BitSets
Refactored code creating classes capable of converting
byte to BitSet to numeric frequency ranges.

e.g. byte[] = {-1} -> 191.325:191.375 THz

Each byte in an array is treated as 8 bits. Meaning an array
of 96 bytes is treated as a BitSet of 768 bits. The bits may
subsequently be converted to a map where the key is the
lower frequency and the value is the upper frequency in a range.

FlexGrid or FixGrid data are treated equally.

Examples

byte[] = {1}       -> BitSet: {0}               -> 191.325  :191.33125
byte[] = {15}      -> Bitset: {0,1,2,3}         -> 191.325  :191.35
byte[] = {-16}     -> BitSet: {4,5,6,7}         -> 191.35   :191.37
byte[] = {-1}      -> BitSet: {0,1,2,3,4,5,6,7} -> 191.325  :191.37
byte[] = {-128}    -> BitSet: {7}               -> 191.36875:191.375
byte[] = {-128, 1} -> BitSet: {7,8}             -> 191.36875:191.38125

Package Overview

The main parts of this package are the classes AvailableGrid and
NumericFrequency. AvailableGrid is used to ease converting
a byte array to either assigned (used) or available frequencies
represented by a BitSet.

AvailableGridFactory provides a way to instantiate an AvailableGrid
object using either available or used frequency grid byte data.

The method getFreqMapFromBitSet in ConvertORToTapiTopology has
been modified and copied to NumericFrequency and is used to
convert a bitset to a Map<Double, Double> (numeric frequency range).

The interface Math is intended to make it easier to unit test the code.
The implementation of Math (i.e. FrequencyMath) simply wraps calls to
GridUtils.getStartFrequencyFromIndex(...) and
GridUtils.getStopFrequencyFromIndex(...).
Since GridUtils depends on global constants, any changes to those
constants may result in unit tests failing. The interface 'Math'
provides a way to test the code using predictable data.

Example usage:

/**
 * This method converts a byte array (e.g. 96 elements) representing
 * frequency ranges into a BitSet (i.e. 768 bits), and finally into
 * a Map<Double, Double> where the key represents the start frequency
 * and the value the end frequency in a range.
 *
 * In short, this example will convert...
 *     byte[] = {-1}
 * ...into...
 *      Map<Double, Double> = 191.325:191.375
 */
public Map<Double, Double> availableMap(byte[] availableByteMap) {

    Available available = new AvailableGrid(availableByteMap);

    // This object will help us convert an instance of 'Available'
    // into a numeric frequency range.
    Numeric numericFrequency = new NumericFrequency(
            GridConstant.START_EDGE_FREQUENCY,
            GridConstant.EFFECTIVE_BITS,
            new FrequencyMath()
    );

    // Convert BitSet data contained in the bitMap object
    // into a frequency range map.
    // Note: If instead we want the assigned frequencies
    //       we could execute:
    //       numericFrequency.assignedFrequency(available):
    return numericFrequency.availableFrequency(available);
}

JIRA: TRNSPRTPCE-802
Change-Id: I578abad37351eb4ead5e8907207f1384df27f09a
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>