New Tapi package dealing with frequency BitSets
Refactored code creating classes capable of converting
byte to bit to numeric frequencies.

e.g. byte[] = {-128} --> 0b10000000 --> 191.325THz

Each byte in an array is treated as 8 bits. Meaning an array
of 96 bytes is treated as a BitSet of 768 bits. The bits may
subsequently be converted to a map where the key is the
lower frequency and the value is the upper frequency in a range.

The package treats each signed byte at "face value". FlexGrid
or FixGrid data are treated equally.

Examples
byte[] = {-16} --> 0b11110000 --> 191.325:191.35625 THz
byte[] = {-1}  --> 0b11111111 --> 191.325:191.375 THz

Package Overview

The main parts of this package are the classes AssignedGrid and
NumericFrequency. AssignedGrid is used to ease converting
a byte array to either assigned (used) or available frequencies
represented by a BitSet.

AssignedGridFactory provides a way to instantiate an AssignedGrid
object using either available or used frequency grid byte data.

The method getFreqMapFromBitSet in ConvertORToTapiTopology has
been modified and copied to NumericFrequency and is used to
convert a bitset to a Map<Double, Double> (numeric frequency range).

The interface Math is intended to make it easier to unit test the code.
The implementation of Math (i.e. FrequencyMath) simply wraps calls to
GridUtils.getStartFrequencyFromIndex(...) and
GridUtils.getStopFrequencyFromIndex(...).
Since GridUtils depends on global constants, any change to those constants
may result in unit tests failing. Hence the interface Math provides a way
to test the code using predictable data.

Example usage adapted from ConvertORToTapiTopology:

/**
 * This method converts a byte array (e.g. 96 elements) representing
 * frequency ranges into a BitSet (i.e. 768 bits), and finally into
 * a Map<Double, Double> where the key represents the start frequency
 * and the value the end frequency in a range.
 */
public Map<Double, Double> available(
        Map<AvailFreqMapsKey, AvailFreqMaps> availFreqMaps) {

    AvailFreqMapsKey availFreqMapsKey =
            new AvailFreqMapsKey(GridConstant.C_BAND);

    // Find available/used frequencies as a byte array (96 elements).
    byte[] frequencyByteMap = availFreqMaps.entrySet().stream()
            .filter(afm -> afm.getKey().equals(availFreqMapsKey))
            .findFirst().orElseThrow().getValue().getFreqMap();

    Numeric numericFrequency = new NumericFrequency(
            GridConstant.START_EDGE_FREQUENCY,
            GridConstant.EFFECTIVE_BITS,
            new FrequencyMath()
    );

    Factory gridFactory = new AssignedGridFactory();

    // The available frequencies are converted from a byte array into
    // an instance of BitMap.
    //
    // In essence, the object is capable of returning available and/or
    // used frequencies as a BitSet (i.e assignedFrequencies() or
    // availableFrequencies()
    //
    // Each value in frequencyByteMap will be converted to 8 bits.
    BitMap bitMap = gridFactory.fromAvailable(freqByteSet);

    return numericFrequency.assignedFrequency(bitMap);
}

JIRA: TRNSPRTPCE-802
Change-Id: I578abad37351eb4ead5e8907207f1384df27f09a
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>