New Tapi package dealing with frequency BitSets
Refactored code creating classes capable of converting
byte to bit to numeric frequencies.

Each byte in an array is treated as 8 bits. Meaning an array
of 96 bytes is treated as a BitSet of 768 bits. The bits may
subsequently be converted to a map where the key is the
lower frequency and the value the upper frequency in a range.

The main parts of this package are the classes BitMapFrequencies and
NumericFrequency. BitMapFrequencies is there to ease converting
a byte array to either assigned (used) or available frequencies
represented by a BitSet.

The method getFreqMapFromBitSet in ConvertORToTapiTopology has
been modified and copied to NumericFrequency.

The interface Math and the implementation is intended to make
it easier to unit test the code. The implemeation of Math simply
wraps a call to GridUtils.getStartFrequencyFromIndex(...) and
GridUtils.getStopFrequencyFromIndex(...).

Example usage adapted from ConvertORToTapiTopology:

/**
 * This method converts a byte array (e.g. 96 elements) representing
 * frequency ranges into a BitSet (i.e. 768 bits), and finally into
 * a Map<Double, Double> where the key represents the start frequency
 * and the value the end frequency.
 */
public Map<Double, Double> available(
        Map<AvailFreqMapsKey, AvailFreqMaps> availFreqMaps) {

    AvailFreqMapsKey availFreqMapsKey = new AvailFreqMapsKey(768);

    // Find available/used frequencies as a byte array (96 elements).
    // Assumed returning a byte array where available frequencies
    // are -1 and used are 0.
    byte[] frequencyByteMap = availFreqMaps.entrySet().stream()
            .filter(afm -> afm.getKey().equals(availFreqMapsKey))
            .findFirst().orElseThrow().getValue().getFreqMap();

    Numeric numericFrequency = new NumericFrequency(
            GridConstant.START_EDGE_FREQUENCY,
            GridConstant.EFFECTIVE_BITS,
            new FrequencyMath()
    );

    Factory frequencyFactory = new FrequencyFactory();

    // The assigned frequencies are converted from a byte array into
    // an instance of BitMap.
    //
    // In essence, the object is capable of returning available and/or
    // used frequencies as a BitSet (i.e assignedFrequencies() or
    // availableFrequencies()
    //
    // The second argument ((byte) 0) instructs the factory
    // to treat values in frequencyByteMap[i] == 0 as "frequency
    // range assigned".
    // All other values are treated as "frequency range available".
    //
    // Each value in frequencyByteMap will be converted to 8 bits.
    BitMap assignedFrequencyBitMap = frequencyFactory
            .assigned(frequencyByteMap, (byte) 0);

    return numericFrequency.availableFrequency(assignedFrequencyBitMap);
}

JIRA: TRNSPRTPCE-802
Change-Id: I578abad37351eb4ead5e8907207f1384df27f09a
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>