New Tapi package dealing to frequency BitSets
Refactored code creating classes capable of converting
byte to bit to numeric frequencies.

Each byte in an array is treated as 8 bits. Meaning an array
of 96 bytes is treated as a BitSet of 768 bits. The bits may
subsequently be converted to a map where the key is the
lower frequncy and the value upper frequency in a range.

The method getFreqMapFromBitSet in ConvertORToTapiTopology has
been modified and copied to NumericFrequency.

Example usage

public Map<Double, Double> available(
        Map<AvailFreqMapsKey, AvailFreqMaps> availFreqMaps) {

    AvailFreqMapsKey availFreqMapsKey = new AvailFreqMapsKey(768);

    byte[] frequencyByteMap = availFreqMaps.entrySet().stream()
            .filter(afm -> afm.getKey().equals(availFreqMapsKey))
            .findFirst().orElseThrow().getValue().getFreqMap();

    Numeric numericFrequency = new NumericFrequency(
            GridConstant.START_EDGE_FREQUENCY,
            GridConstant.EFFECTIVE_BITS,
            new FrequencyMath()
    );

    Factory frequencyFactory = new FrequencyFactory();

    BitMap assignedFrequencyBitMap = frequencyFactory
            .assigned(frequencyByteMap, (byte) 0);

    // Using numericFrequency.assignedFrequency(assignedFrequencyBitMap)
    // will return the used frequencies instead.
    return numericFrequency.availableFrequency(assignedFrequencyBitMap);
}

Change-Id: I578abad37351eb4ead5e8907207f1384df27f09a
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>