Error in TAPI creating frequency ranges
When an OpenROADM service is converted to TAPI, an exception is thrown
due to array index out of bounds. An byte array (size 96) was used when
the code was actually expecting size 768 (i.e. 96 x 8), when trying
to create a Map<Double, Double> with frequency ranges.

The use of method getFreqMapFromBitSet in ConvertORToTapiTopology
has been replaced with the package
org.opendaylight.transportpce.tapi.frequency.

A refactored version of getFreqMapFromBitSet can be found in the class
org.opendaylight.transportpce.tapi.frequency.NumericFrequency.

Change-Id: I423507decaae38b44d9f6968882179c3ec1b11d4
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>