New package tapi/frequency
Frequency
------------------

The core of this package is the interface Frequency and
the implementing class TeraHertz. In essense, TeraHertz is nothing
more than a class capable of converting a frequency into different
data types. i.e. Uint64 as Hz and BigDecimal as THz.

Internally, frequency is oftentimes represented by a Double in THz,
while TAPI typically communicates frequency in Hz as Uint64. The
TeraHertz class is intended to be able to convert between these
two formats without rounding errors.

The TeraHertzFactory may be used to create instances of Frequency by
using center frequency and width, or a start frequncy, bitNumber
and granularity.

Range
------------------

Oftentimes we need a frequency range as opposed to just a frequency.
The package frequency/range provides the interface Range and its
implementation SortedRange. The class may be used to create ranges
as a collection of Frequency pairs. Each pair uses the key as
the lower and value as the upper frequency in the range.

The interface Range, just like the interface Frequency, provides
methods retrieving ranges in different formats/data types
(e.g. Uint64, Double or Frequency objects). It's also possible to
add a range to another range creating a collection of ranges.

The interface RangeFactory and its implementation FrequencyRangeFactory
may be used to create a range using a center frequency in THz and a
bandwidth in GHz. The method effectiveRange() may be used to create
a single range using start frequency, granularity and effective bits.
Useful in those cases where the entire spectrum is "available" or
"assigned".

Examples
------------------

    System.out.println("===== Frequency datatypes =====");
    Frequency hertz = new TeraHertz(193.35);
    System.out.printf("Frequency : %sTHz\n", hertz.teraHertz());
    // Frequency : 193.35THz

    System.out.printf("Frequency : %sHz\n", hertz.hertz());
    // Frequency : 193350000000000Hz

    System.out.println("\n===== Range =====");
    Range range = new SortedRange();
    Frequency lowerFrequency = new TeraHertz(191.40);
    Frequency upperFrequency = new TeraHertz(
            BigDecimal.valueOf(191.45)
    );
    range.add(lowerFrequency, upperFrequency);
    range.add(191.35, 191.40);

    Map<Frequency, Frequency> rangesAsFrequency = range.ranges();
    Map<Double, Double> rangesAsDouble = range.rangesAsDouble();
    Map<Uint64, Uint64> rangesAsUint64 = range.rangesAsUint64();

    System.out.printf("Frequency objects: %s\n", rangesAsFrequency);
    // Frequency objects: {191.35=191.4, 191.4=191.45}

    System.out.printf("Double objects:    %s\n", rangesAsDouble);
    // Double objects:    {191.35=191.4, 191.4=191.45}

    System.out.printf("Uint64 objects:    %s\n", rangesAsUint64);
    // Uint64 objects:    {191350000000000=191400000000000,
    //                     191400000000000=191450000000000}

    System.out.println("\n===== RangeFactory =====");
    RangeFactory rangeFactory = new FrequencyRangeFactory();
    Range center = rangeFactory.range(191.50, 50.0);
    System.out.println(center.ranges());
    // {191.475=191.525}

    System.out.println("\n===== Add a range to another range =====");
    center.add(range);
    System.out.println(center.ranges());
    // {191.35=191.4, 191.4=191.45, 191.475=191.525}

Change-Id: Ic03aef368f2a8d56d39a6e7756434e27b4525e8b
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>