New package tapi/frequency
The OpenROADM models typically work with THz, while TAPI works with Hz.
The TAPI implementation therefore need to convert THz to Hz when
converting an OpenROADM service to a TAPI service (as an example).

This package aims at simplifying the process by executing the required
conversion math indirectly, as opposed to directly.

Frequency
------------------

The core of this package is the interface Frequency and
the implementing class TeraHertz. In essense, TeraHertz is nothing
more than a class capable of expressing a frequency using different
units (i.e. THz or Hz).

Range
------------------

The core of this package is the interface Range and its
implementation SortedRange. The class may be used to create ranges
as a collection of Frequency pairs. Each pair uses the key as
the lower and value as the upper frequency in the range.

The interface Range, just like the interface Frequency, provides
methods retrieving ranges in different units (i.e. THz or Hz).

Example #1 - Converting THz to Hz
---------------------------------

    Frequency freq = new TeraHertz(193.35);
    System.out.printf("Frequency : %sTHz\n", freq.teraHertz());
    // Frequency : 193.35THz

    System.out.printf("Frequency : %sHz\n", freq.hertz());
    // Frequency : 193350000000000Hz

EXAMPLE #2 - Convert frequency range
------------------------------------

    Range range = new SortedRange();
    range.add(new TeraHertz(191.40), new TeraHertz(191.45));
    range.add(191.35, 191.40);

    Map<Frequency, Frequency> rangesAsFrequency = range.ranges();
    Map<Double, Double> rangesAsTHz = range.asTeraHertz();
    Map<Uint64, Uint64> rangesAsHz  = range.asHertz();

    System.out.printf("Frequency objects: %s\n", rangesAsFrequency);
    // Frequency objects: {191.35=191.4, 191.4=191.45}

    System.out.printf("Double objects:    %s\n", rangesAsTHz);
    // Double objects:    {191.35=191.4, 191.4=191.45}

    System.out.printf("Uint64 objects:    %s\n", rangesAsHz);
    // Uint64 objects:    {191350000000000=191400000000000,
    //                     191400000000000=191450000000000}

EXAMPLE #3 - Rangefactory
-------------------------

    RangeFactory rangeFactory = new FrequencyRangeFactory();
    Range center = rangeFactory.range(191.50, 50.0);
    System.out.println(center.ranges());
    // {191.475=191.525}

    center.add(range);
    System.out.println(center.ranges());
    // {191.35=191.4, 191.4=191.45, 191.475=191.525}

JIRA: TRNSPRTPCE-865
Change-Id: Ic03aef368f2a8d56d39a6e7756434e27b4525e8b
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>