TAPI resolving an issue converting THz to Hz
Uses the package tapi/frequency to solve an issue where
TAPI returns available frequency as THz (e.g. 191.35)
as opposed to Hz (e.g. 191350000000000).

The primary change is in the method populateNepsForRdmNode
int the class ConvertORTopoToTapiFullTopo. While the rest of
the method converts THz to Hz, these lines does not:

    case DEGREETXRXTTP:
        usedFreqMap = tapiFactory.getTTPUsedFreqMap(tp).ranges();
        availableFreqMap =
            tapiFactory.getTTPAvailableFreqMap(tp).ranges();
        break;

The rest of the changes are mostly ripple effects.

Change-Id: Iadadfc16cfdad583328cc2c78aa448a08e08b34e
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>