TAPI resolving an issue frequency as THz
Uses the package tapi/frequency to solve an issue where
TAPI return available frequency as THz as opposed to Hz.

The primary change is in the method populateNepsForRdmNode
int the class ConvertORTopoToTapiFullTopo.

While the rest of the method converts frequency to Hz, these
lines did not. The rest of the changes are ripple effects.

    case DEGREETXRXTTP:
        usedFreqMap = tapiFactory.getTTPUsedFreqMap(tp).ranges();
        availableFreqMap =
            tapiFactory.getTTPAvailableFreqMap(tp).ranges();
        break;

Change-Id: Iadadfc16cfdad583328cc2c78aa448a08e08b34e