TAPI - Refactor code from THz to Frequency
Refactoring TAPI code to use implementations of Frequency
as opposed to primitive values. This lowers the possibility
of rounding errors and ensures not mixing up THz with Hz
by accident.

Converting THz to Hz is achieved by calling a method, resulting in
conversion math being executed indirectly as opposed to directly.

Change-Id: I4553e08816a140052e60ed5ef61ca7440c5a02d9
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>