TAPI refactor code from primitives to Frequency
Refactoring TAPI code to use implementations of Frequency
as opposed to primitive values. This lowers the possibility
of rounding errors and ensures not mixing up THz with Hz
by accident.

Also, converting THz to Hz can be done by calling a method as
opposed to doing math, avoiding potentially introducing
rounding errors in the process.

Change-Id: I4553e08816a140052e60ed5ef61ca7440c5a02d9
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>