Correct bugs handling spectrum constraint
When defining customer profiles for spectrum filling (using controller-
customization model defined in OpenROADM), the dedicated-signal-
badwidth multiple is not consider by the controller to set the width
of the ROADM connections. This patch corrects the bug.
- nmc-bandwidth set in ModelMappingUtils according to the min and max
frequencies of path description, if those are provided.
-In PostAlgoValidator, feed operational-mode and osnr in PceResult
(API was completed with additional parameters that were not fed)

JIRA: TRNSPRTPCE-826
Signed-off-by: orenais <olivier.renais@orange.com>
Change-Id: If376072dc122a4ed34f7e16cd2bbda5c9e2b154b