Fix spectrum assignment computation
The loop over available spectrum used BitSet.size() as starting value.
But size() returns the number of bits of space actually in use by the
object, which is implementation dependent. This means the loop
sometimes uses more iterations than necessary but worse, it could
happen that spectrum assignment fails even if available spectrum exists
since the size could be smaller than the smallest index of available
spectrum.

Change to use BitSet.length() instead, which returns the index of the
highest set bit in the BitSet plus one.

JIRA: TRNSPRTPCE-832

Signed-off-by: Jonas Mårtensson <jonas.martensson@smartoptics.com>
Change-Id: I1b6794b19929f1ece183de1ccdea2ec05411e126