New package simplifying picking service frequency
The intention behind the package is to simplify selecting
a frequency range for a new service.

Choosing a frequency range depends on these criteria:

1) Is the "customer" limited to a certain range?
2) The available range on the nodes for the service.
3) The service format (type).
4) The requested frequency range presented in the service
   create request (i.e. API body).
5) The requested center frequency presented in the
   service create request (i.e. API body).

The class FrequencySelectionFactory brings all the pieces together
in this package into solving the requirements above. There is an
example showing how the package may be used further down below.

They idea behind the package was to create easily testable code
with a minimum amount of dependencies with other parts of the
code base.

Also, I wanted the code to make as few assumptions as possible and
instead rely on being "fed" the facts from the outside.
As an example, the code is asking for frequency granularity as opposed
to reaching out to a global constant "behind the scenes". Instead,
the global constant can be "fed" into the class or method where
it's needed.

Finally, I wanted to change as little code as possible in existing
classes. Moving the code and logic into separate classes really
helped in that regard.

Interfaces Interval and Collection
----------------------------------

A few key components in this package are the interfaces Interval and
Collection found in the sub package interval. The Interval interface
simply defines a frequency interval as a start and stop frequency.
Regardless of whether or not we're dealing with frequency-range or
frequency-slots inputs, we can 'model' the information as a frequency
interval.
The implementing class also contains methods allowing two
intervals to be compared and allows us to easily remove duplicate
information should an interval be added to a List or Set requiring
unique values.

The Collection interface defines methods for a collection of intervals.
Meaning we can add one or more intervals to the collection. We may then
execute various methods on the collection suitable for calculating
the intersection or subset between any number of frequency intervals
in the collection. The main purpose of the Collection interface is to
be able to compute intersections and subset between frequency ranges
and BitSets.

Since the API input allows the user to specify input in two different
locations, it made sense to create a collection handling these
frequency intervals and let the implementing class deal with
the complexity involved handling potentially conflicting inputs.

Having the client limit the scope (frequency range) could be considered
as calculating the intersection between input and available frequency
range.
Having the client specify an exact center frequency and slot
width could be considered calculation the subset between input and
available frequency. Meaning the entire range defined by center
frequency and slot width needs to be a real subset to available
frequencies.

The class EntireSpectrum is a special kind of Collection. In essence
it's intended to represent a NULL-value. Assuming the API request body
doesn't contain any frequency information, by using the class
EntireSpectrum while calculating the intersection between
the "Entire frequency range" and the available BitSet, we can ensure
the entire available spectrum may be used for a service. i.e. we're not
restricting any of the available range(s) from being used. Furthermore,
by using an object representing NULL as opposed to an actual NULL
value, we can eliminate the need for some NULL-checks and subsequently
avoiding NullPointerException's in the process.

Example usage
-------------

/**
 * Determine the available spectrum for the new service.
 * A frequence range from a service may be picked from
 * the output of this method.
 *
 * @param input API input.
 * @param availableSpectrumOnNodes
 *             The available spectrum on the nodes
 *             selected for the new service.
 * @param customerSpectrumConstraint
 *             The customer may be restricted to this
 *             spectrum.
 * @return The spectrum that is available for the new
 *         service.
 */
BitSet availableSpectrum(
        PathComputationRequestInput input,
        BitSet availableSpectrumOnNodes,
        BitSet customerSpectrumConstraint) {

    ClientInput clientInput = new ApiClientInput(
            input,
            new FrequencyIntervalFactory(768, 6.25),
            new FrequencySpectrum(
                    new SpectrumIndex(
                            191.325,
                            6.25,
                            768
                    )
            ),
            new ServiceFrequency(),
            6.25
    );

    Select frequencyFactory =
            new FrequencySelectionFactory(768);

    BitSet availableSpectrum =
            frequencyFactory.availableFrequencyRange(
                    clientInput,
                    customerSpectrumConstraint,
                    availableSpectrumOnNodes
    );

    return availableSpectrum;
}

Change-Id: I565198a5906ac600ada204b100f41eb9fe7aeb82
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>